requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.3.0
joblib==1.0.0
Markdown==3.3.3
nltk==3.5
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.5
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
textblob==0.15.3
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
xmltodict==0.12.0
zipp==3.4.0
